third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-memory.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function dumpInnerHtml()
{
    frame = document.documentElement.appendChild(document.createElement("iframe"));
    frame.src = "resources/iframe.html";
    frame.onload = function()
    {
        var location = frame.contentWindow.location;
        frame.remove();
        memory = console.__lookupGetter__("memory").call(location);
        console.log(memory.constructor.constructor("return document.querySelector(\"#mydiv\").textContent")());
    }
}

function test()
{
    InspectorTest.sendCommandOrDie("Runtime.enable", {});
    InspectorTest.eventHandler["Runtime.consoleAPICalled"] = dumpMessageAndCompleTest;
    InspectorTest.sendCommand("Runtime.evaluate", { "expression": "dumpInnerHtml()" });

    function dumpMessageAndCompleTest(result)
    {
        InspectorTest.log("=== Dump console message ===");
        InspectorTest.log(result.params.args[0].value);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">Tests that console.memory works correct.</body>
<div id="mydiv">div from page</div>
</html>
