third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-memory-expected.txt

CONSOLE MESSAGE: line 15: div from page
Tests that console.memory works correct.
div from page
=== Dump console message ===
div from page
